Changes to reduce warnings using ia16-unknown-elks-gcc
- Reduce warnings using ia16-unknown-elks-gcc
- Elimination of inline assembly within several C functions.
- Fixed prototypes of several functions.
- Simplified file arch/i86/drivers/char/bell.c.

Code and data size was unchanged.

+++ elks/fs/minix/dir.c
@@ -16,7 +16,7 @@
 
 #include <arch/segment.h>
 
-static size_t minix_dir_read(struct inode *inode, struct file *filp, char *buf, int count)
+static size_t minix_dir_read(struct inode *inode,struct file *filp,char *buf,size_t count)
 {
     return -EISDIR;
 }
